+++ front-end/src/components/DashBoardProjects.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import DashBoardBoxes from './DashBoardBoxes'
+
 import {makeStyles} from '@material-ui/core/styles'
 import Styled from "styled-components";
 
@@ -38,6 +39,7 @@ export default function DashBoardProjects(){
     const classes = boxDiv()
     return(
         <div>
+        <div>
              <TitleContainer>
                 <Title>Dashboard</Title>
                 <Paragraph>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</Paragraph>
@@ -46,6 +48,7 @@ export default function DashBoardProjects(){
             <div>
                 <DashBoardBoxes />
             </div>
+         </div>  
         </div>
     )
 }
