+++ front-end/src/components/Dashboard.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import LeftNav from "./leftnav.js";
+import NavBarDashboard from "./NavBarDashboard.js";
 
 
 import Styled from "styled-components";
@@ -13,9 +14,11 @@ const Background = Styled.div `
 
 export default  function Dashboard(){
     return(
+        <div>
+        <NavBarDashboard />
         <Background>
             {/* Add your component here */}
              <LeftNav />
         </Background>
-     
+        </div>
    ) }
